backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy[asyncio]
asyncpg
pydantic[email]
pydantic-settings
python-dotenv
passlib[bcrypt]
bcrypt
python-multipart
pytest
pytest-asyncio
httpx
anyio
